projects/ERP_Connect_CognitION/requirements.txt
# ERP Connect CogitION v2.0 Requirements
# Python 3.11 recommended

# Core ML/NLP
transformers>=4.35.0
datasets>=2.14.0
peft>=0.7.0
sentence-transformers>=2.2.2
accelerate>=0.24.0

# Vector search
faiss-cpu>=1.7.4

# Evaluation metrics
rouge-score>=0.1.2
nltk>=3.8.1
scikit-learn>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.1.0
pyyaml>=6.0.1

# Web interface
gradio>=4.0.0

# Utilities
tqdm>=4.66.0
